refactor to keep consistencies in files
