--- apps/web/app/future/enterprise/layout.tsx
@@ -1,5 +1,0 @@
-import { WithLayout } from "app/layoutHOC";
-
-import { getLayout } from "@calcom/features/MainLayoutAppDir";
-
-export default WithLayout({ getLayout })<"L">;
