+++ apps/web/app/future/event-types/page.tsx
@@ -1,5 +1,8 @@
 import EventTypes from "@pages/event-types";
 import { _generateMetadata } from "app/_utils";
+import { WithLayout } from "app/layoutHOC";
+
+import { getLayout } from "@calcom/features/MainLayoutAppDir";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -7,4 +10,4 @@ export const generateMetadata = async () =>
     (t) => t("event_types_page_subtitle")
   );
 
-export default EventTypes;
+export default WithLayout({ getLayout, Page: EventTypes })<"P">;
