+++ apps/web/app/future/reschedule/[uid]/page.tsx
@@ -1,11 +1,7 @@
-import OldPage, { getServerSideProps as _getServerSideProps } from "@pages/reschedule/[uid]";
+import OldPage, { getServerSideProps } from "@pages/reschedule/[uid]";
 import { withAppDirSsr } from "app/WithAppDirSsr";
-import type { SearchParams } from "app/_types";
 import { _generateMetadata } from "app/_utils";
-import type { Params } from "next/dist/shared/lib/router/utils/route-matcher";
-import { headers, cookies } from "next/headers";
-
-import { buildLegacyCtx } from "@lib/buildLegacyCtx";
+import { WithLayout } from "app/layoutHOC";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -13,19 +9,6 @@ export const generateMetadata = async () =>
     () => ""
   );
 
-type PageProps = Readonly<{
-  params: Params;
-  searchParams: SearchParams;
-}>;
+const getData = withAppDirSsr(getServerSideProps);
 
-const getData = withAppDirSsr(_getServerSideProps);
-
-const Page = async ({ params, searchParams }: PageProps) => {
-  const legacyCtx = buildLegacyCtx(headers(), cookies(), params, searchParams);
-
-  await getData(legacyCtx);
-
-  return <OldPage />;
-};
-
-export default Page;
+export default WithLayout({ getLayout: null, getData, Page: OldPage });
